Improve images and surfaces management and add PNG loading support
